Modified python graded example
