+++ exampleCourse/courseInstances/Sp15/assessments/hw06-demoPythonAutograder/infoAssessment.json
@@ -41,7 +41,7 @@
                 {"id": "demo/autograder/python/plots",          "points": 1, "maxPoints": 5},
                 {"id": "demo/autograder/python/random",         "points": 1, "maxPoints": 5},
                 {"id": "demo/autograder/ansiOutput",            "points": 2, "maxPoints": 4},
-                {"id": "demo/autograder/python/dragDropCode_simple",           "points": 1, "maxPoints": 2},
+                {"id": "demo/autograder/python/orderBlocks_randomParams",      "points": 1, "maxPoints": 2},
                 {"id": "demo/autograder/python/dragDropCode_addNumpy",         "points": 1, "maxPoints": 2}
             ]
         }
