--- exampleCourse/questions/demo/autograder/python/dragDropCode_simple/question.html
@@ -1,20 +1,0 @@
-<pl-question-panel>
-  <p>
-    Write a Python function that returns an even number.
-  </p>
-
-</pl-question-panel>
-
-<pl-hide-in-panel answer="true">
-<pl-order-blocks answers-name="coding" external-grader="true">
-  <pl-answer>def return_even():</pl-answer>
-  <pl-answer>return 1</pl-answer>
-  <pl-answer>return 2</pl-answer>
-  <pl-answer>return 4</pl-answer>
-</pl-order-blocks>
-</pl-hide-in-panel>
-
-<pl-submission-panel>
-  <pl-external-grader-results></pl-external-grader-results>
-  <pl-file-preview></pl-file-preview>
-</pl-submission-panel>
