--- exampleCourse/questions/demo/autograder/python/dragDropCode_simple/server.py
@@ -1,5 +1,0 @@
-def generate(data):
-    data['params']['names_for_user'] = []
-    data['params']['names_from_user'] = [
-        {'name': 'return_even', 'description': 'a function that returns a hard-coded even number', 'type': 'python function'}
-    ]
