+++ exampleCourse/questions/demo/autograder/python/orderBlocks_randomParams/info.json
@@ -1,6 +1,6 @@
 {
     "uuid": "3551731f-b6b4-4afc-b1f8-a8a4f7f73bcd",
-    "title": "pl-order-blocks: Ordering lines of autograded python code",
+    "title": "pl-order-blocks: Externally graded python code with randomized parameters",
     "topic": "Autograder",
     "tags": ["code", "v3"],
     "type": "v3",
@@ -10,6 +10,6 @@
         "enabled": true,
         "image": "prairielearn/grader-python",
         "entrypoint": "/python_autograder/run.sh",
-        "timeout": 5
+        "timeout": 20
     }
 }
