+++ exampleCourse/questions/demo/autograder/python/orderBlocks_randomParams/question.html
@@ -1,0 +1,22 @@
+<pl-question-panel>
+  <p>
+    Write a Python function that returns an {{params.type}} number.
+  </p>
+
+</pl-question-panel>
+
+<pl-hide-in-panel answer="true">
+<pl-order-blocks answers-name="coding" grading-method="external">
+  <pl-answer>def return_number():</pl-answer>
+  <pl-answer>{{params.option1}}</pl-answer>
+  <pl-answer>{{params.option2}}</pl-answer>
+  <pl-answer>{{params.option3}}</pl-answer>
+  <pl-answer>{{params.option4}}</pl-answer>
+  <pl-answer>return a</pl-answer>
+</pl-order-blocks>
+</pl-hide-in-panel>
+
+<pl-submission-panel>
+  <pl-external-grader-results></pl-external-grader-results>
+  <pl-file-preview></pl-file-preview>
+</pl-submission-panel>
