+++ exampleCourse/questions/demo/autograder/python/orderBlocks_randomParams/tests/setup_code.py
@@ -1,2 +1,4 @@
 # Anything in this file will be run before the student's code
 # Use it to generate anything needed for the solution
+
+number_type = data['params']['type']
